+++ CHANGELOG.md
@@ -4,6 +4,13 @@ As of version 2.0.0, all notable changes to HTML Minifier Next are documented in
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.1] - 2025-10-29
+
+### Internal
+
+- Restructured tests
+- Added CSS and JS minification tests
+
 ## [3.2.0] - 2025-10-28
 
 ### Added